Add language selection to narrow detection scope
So far all the language data was used when detecting the language of an
input. If the corpus is large enough, this is definitely feasible. The
input, however, is sometimes short consisting only of few words.

To avoid false detections the constructor should be used to narrow the
scope of possible languages. An instance variable @selection is used.

In the method for processing the text this variable is used. Unless the
symbol :all was passed, the intersection of all the possible languages
and the selection is created.

A simple unit test was added to test this behavior. In this case the
input consists of three German articles. Without the language selection
Russian is detected as language. With the language selection, German is
detected.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>